requirements_old.txt
about-time==4.2.1
absl-py==2.1.0
alive-progress==3.1.5
astunparse==1.6.3
autograd==1.6.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
cma==3.2.2
contourpy==1.2.0
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
dm-tree==0.1.8
flatbuffers==24.3.7
fonttools==4.50.0
future==1.0.0
gast==0.5.4
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grapheme==0.6.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
intel_tensorflow==2.14.0
intel_tensorflow_avx512==2.14.0
kagglehub==0.2.0
keras==2.14.0
keras-core==0.1.7
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pymoo==0.6.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rich==13.7.1
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow-addons==0.23.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.36.0
tensorflow-text==2.16.1
termcolor==2.4.0
tqdm==4.66.2
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.14.1
